Align configuration contracts with Mermaid schema

+++ src/mermaiden/diagrams/mindmap/configuration.py
@@ -6,6 +6,6 @@ from ..domain import MermaidDiagramConfiguration
 
 class MindmapDiagramConfiguration(MermaidDiagramConfiguration):
     use_max_width: bool = True
-    padding: int = 10
-    max_node_width: int = 200
+    padding: float = 10
+    max_node_width: float = 200
     layout_algorithm: str = Field(default="cose-bilkent", pattern=Text.pattern, description=Text.description)
